fix(qwen3.5): keep GDN scalar params in model dtype; default shared core off
Merging gfx906/main introduced two regressions on GGUF Qwen3.5 (gfx906):

1. The shared GatedDeltaNetAttention base forced A_log to fp32 while the
   model is built under set_default_torch_dtype(fp16); the old specialized
   base created it in model dtype. The gfx906 packed-decode kernels read
   A_log with fp16 element size, so the fp32 tensor made them misread the
   dt decay rate: decode output collapsed to a constant and the recurrent
   state grew unbounded to inf within two steps (duplicate-token output
   'Let!!!!' at 0.66-2.4 chunks/s vs 20.5 healthy). Drop the forced fp32
   so A_log follows the model dtype again. Verified: layer-0 stage stats
   match the pre-merge tree bit-for-bit on prefill; probe recovers
   20.7 chunks/s with coherent output.

2. The shared core (VLLM_QWEN35_SHARED_GDN_CORE=1) still produces
   deterministic token-salad near the tail of long generations even with
   the A_log fix (reproduced twice, identical corrupted tail), while the
   specialized Qwen3NextGatedDeltaNet core is fully coherent. Default the
   flag to False until the shared core passes full generation checks;
   VLLM_QWEN35_SHARED_GDN_CORE=1 still opts in.

Smoke-tested:
45s streaming probe 20.66 chunks/s, 3802 chars, head and tail coherent.

(cherry picked from commit 8dbbdf8c16c2421e49b8396149aed10953655ce0)

+++ vllm/model_executor/layers/mamba/gdn_linear_attn.py
@@ -363,8 +363,13 @@ class GatedDeltaNetAttention(nn.Module, MambaBase):
         )
 
         self.dt_bias = nn.Parameter(torch.ones(self.num_v_heads // self.tp_size))
+        # A_log/dt_bias must follow the model dtype (fp16 for GGUF Qwen3.5).
+        # The gfx906 packed-decode kernels read these scalars with fp16
+        # element size; a forced fp32 A_log makes them misread the decay
+        # rate, so the recurrent state grows unbounded (-> inf) and decode
+        # output degenerates.
         self.A_log = nn.Parameter(
-            torch.empty(divide(self.num_v_heads, self.tp_size), dtype=torch.float32)
+            torch.empty(divide(self.num_v_heads, self.tp_size))
         )
 
         set_weight_attrs(self.A_log, {"weight_loader": sharded_weight_loader(0)})
